Don't colorize piped output
Concept behind using the posix_isatty() function to determine whether or
not to colorize output. This probably needs refactoring somehow - for
one, this assumes that checking the $out stream is sufficient, but the
checking should be done on a per-stream basis. Also, it would be nice to
be able to pass $colorize as a variable (for example, in a --porcelain
option of a script), but since this function is already accepting so
many different types of output, it would be very hard to allow that
without breaking existing implementations.