Adds support for SHELL_PIPE Environment Variable.
The isPiped() method now looks for the SHELL_PIPE environment variable.
If found it returns true/false based on the value of SHELL_PIPE. If the
variable is absent existing behaviour to detect a tty remains.

Valid values are boolean strings like 1, 0, yes, no, etc.

To make the CLI ignore that the current TTY is a PIPE and thus enable
ASCII formatting use, SHELL_PIPE=0.