Properly handle line breaks in column value
Before this code tried to replace newlines with empty spaces, but it
didn't always work because the new value would only use the replaced
string if additional wrapping needed to be done due to size constraints.
This updates the code to properly handle new lines in all content by
creating an extra row for each one, using the same logic that already
existed for wrapping longer content.

See https://github.com/wp-cli/entity-command/issues/262 for an example
of the currently broken behavior that this PR fixes