Fix removal of trailing tab / whitespace in tabular table (#184)
Previously the final table output was trimmed with the intention of
removing the last newline. This had an unintended side effect of
removing the tab characters from empty columns in the specific case
where the last row had one or more empty columns at the end of the row.
This makes sure we are only removing the newline character as intended.

Add a new unit test to verify this behavior as well.